Add APIGatewayV2WebSocketEvent/Response as renames of APIGatewayV2Proxy* classes
